EC-13

+++ ElectronicCard/EC.BusinessLogic/Services/Implementation/PhotoService.cs
@@ -17,29 +17,50 @@ namespace EC.BusinessLogic.Services.Implementation
 
         public void CreatePhoto(Photo photo)
         {
+            if (photo == null)
+            {
+                return;
+            }
+
             _photoRepository.Create(photo);
         }
 
         public void UpdatePhoto(Photo photo)
         {
+            if (photo == null)
+            {
+                return;
+            }
+
             _photoRepository.Update(photo);
         }
 
         public void DeletePhoto(int? id)
         {
+            if (id == null || id <= 0)
+            {
+                return;
+            }
+
             _photoRepository.Delete(id);
         }
 
         public Photo GetById(int? id)
         {
-            return id == null ? null : _photoRepository.GetById(id);
+            return id == null || id <= 0? null : _photoRepository.GetById(id);
         }
 
         public Photo GetUserPhoto(string login)
         {
+            if (login == null)
+            {
+                return null;
+            }
+
             var user = _userRepository.GetUserByLogin(login);
 
             return user == null ? null : _photoRepository.GetUserPhoto(user.Id);
+
         }
     }
 }
